third_party/blink/web_tests/platform/mac-mac15-arm64/compositing/webgl/webgl-background-color-expected.txt

ALERT: No WebGL context found
CONSOLE ERROR: Uncaught TypeError: Cannot read properties of null (reading 'viewport')
